added confirmation when delete project

+++ app/assets/javascripts/models/projects.coffee
@@ -13,10 +13,11 @@ angular.module('taskApp').factory 'projects', [
       )
 
     o.destroy = (project)->
-      $http.delete('/api/projects/' + project.id)
-        .success(
-          o.projects.splice(o.projects.indexOf(project),1)
-      )
+      if confirm("You are sure?")
+        $http.delete('/api/projects/' + project.id)
+          .success(
+            o.projects.splice(o.projects.indexOf(project),1)
+        )
 
 
     o
